dialogue working WELL
I made a class that stores dialogue lines in an object that's basically an array of strings. The code can read this array and type it out all cool like. I'm gonna add more info that the class can hold like maybe path to a picture of the guy who's talking or what their "color" is.